[NOT FOR PRODUCTION] Added wrapping method for sale.affiliate.request default values
Since in Odoo 12.0, various fields using the odoo.http.request object have been set as required,
when upgrading the module, the default value is recomputed for an empty recordset (see "update_db_notnull" method).

However, since the default values use the odoo.http.request object, which is unbound outside of
a HTTP request context, this leads to a failure of module upgrading.

As such, we add a fallback try-except block, which detects if the object is unbound (RuntimeError),
and assigns a dummy string if so.